[bug][bsp] Stack overwritten by D-cache.
redmine: #id, REDMINE-id

ext-redmine: #5609, REDMINE-5609

[Description in detail]
Typically, the stack of 'mpu_config' is in SRAM , and its cache is disabled by default.
But during the 'mpu_config' execution, its cache is enabled when the MPU is disabled.
As a result, the stack of 'mpu_config' is cached in the D-cache, and it will overwrite SRAM when the next D-cache flushing is triggered.

Change-Id: I1bba83e1fd19ea9750e5e857157ccbd637e0aa9f